Extract for events sub matrix, convert to wave values, write Data.mat
EEGLearn library supports a special format of the input data.
For all csv files and events 1.3 secs of data for all 56 channels are
extracted, are converted to average wave data in
the form [theta_1, ..., theta_56, alpha_1, ..., beta_1, ..., beta_56]
using EEGrunt and concatenated in the file Data.mat as variable featMat.
Also trainlabels are added to this file.